+++ src/Cache.js
@@ -16,25 +16,26 @@ export class Cache {
     });
   }
 
-  getAnyData(queryUniqueId) {
+  getAnyData(queryUniqueId, _root) {
     return {
       action: "clean",
       source: {
         _id: `${this._id}${queryUniqueId ? `-${queryUniqueId}` : ""}`,
-        _queryId: queryUniqueId
+        _queryId: queryUniqueId,
+        _root
       }
     };
   }
 
-  clean(query) {
+  clean(query, _root) {
     if (query) {
       const queryIdentifier = queryId(query);
       delete this._cachedData[queryIdentifier];
       this._eventEmitter.emit(cleanCacheEventName(query), query);
-      this._eventEmitter.emit(CLEAN_ANY_EVENT_NAME, this.getAnyData(queryIdentifier));
+      this._eventEmitter.emit(CLEAN_ANY_EVENT_NAME, this.getAnyData(queryIdentifier, _root));
     } else {
       this._reset();
-      this._eventEmitter.emit(CLEAN_ANY_EVENT_NAME, this.getAnyData());
+      this._eventEmitter.emit(CLEAN_ANY_EVENT_NAME, this.getAnyData(null, _root));
     }
   }