+++ test/Origin.queries.js
@@ -61,13 +61,13 @@ test.describe("Origin queries", () => {
     });
 
     test.describe("whith query", () => {
-      const FOO_QUERY = { foo: "foo", foo2: "foo-2" };
+      const FOO_CUSTOM_QUERY = { foo: "foo", foo2: "foo-2" };
       test.it("should pass the query value to the method", () => {
         return testOrigin
-          .query(FOO_QUERY)
+          .query(FOO_CUSTOM_QUERY)
           .create()
           .then(() => {
-            return test.expect(spys.create.getCall(0).args[0]).to.deep.equal(FOO_QUERY);
+            return test.expect(spys.create.getCall(0).args[0]).to.deep.equal(FOO_CUSTOM_QUERY);
           });
       });
 
@@ -170,6 +170,12 @@ test.describe("Origin queries", () => {
       });
     });
 
+    test.describe("when created", () => {
+      test.it("public customQueries property should be available", () => {
+        test.expect(testOrigin.customQueries.byId(FOO_ID)).to.deep.equal(FOO_CUSTOM_QUERY_RESULT);
+      });
+    });
+
     test.describe("when testing", () => {
       test.it("should be available at the test.customQueries object", () => {
         test
@@ -177,6 +183,7 @@ test.describe("Origin queries", () => {
           .to.deep.equal(FOO_CUSTOM_QUERY_RESULT);
       });
     });
+
     test.describe("when instance is not queried", () => {
       test.it("should apply the result of the custom query function as query", () => {
         return testOrigin
@@ -191,6 +198,24 @@ test.describe("Origin queries", () => {
     });
 
     test.describe("when instance is queried", () => {
+      test.it("customQueries property should be still available", () => {
+        test.expect(testOrigin.query(FOO_QUERY).customQueries.byId).to.not.be.undefined();
+      });
+
+      test.it(
+        "should have available _id property of root instance in the _root._id property",
+        () => {
+          test.expect(testOrigin.query(FOO_QUERY)._root._id).to.equal(testOrigin._id);
+        }
+      );
+
+      test.it(
+        "should have available all properties of root instance in the _root property",
+        () => {
+          test.expect(testOrigin.query(FOO_QUERY)._root).to.equal(testOrigin);
+        }
+      );
+
       test.it("custom query should be still available", () => {
         test.expect(testOrigin.query(FOO_QUERY).byId).to.not.be.undefined();
       });
@@ -230,6 +255,7 @@ test.describe("Origin queries", () => {
         })
       });
     });
+
     test.it("should apply the result of the custom queries functions results extended", () => {
       return testOrigin
         .byId(FOO_ID)
@@ -245,6 +271,10 @@ test.describe("Origin queries", () => {
             }
           });
         });
+    });
+
+    test.it("should have available all properties of root instance in the _root property", () => {
+      test.expect(testOrigin.byId(FOO_ID).byName(FOO_NAME)._root).to.equal(testOrigin);
     });
   });
 });